bot/requirements-telegram.txt
python-telegram-bot==21.5
python-dotenv
requests
pytz
psutil
aiohttp
gtts
sqlalchemy==2.0.7
requests-cache